separate models tests
